feat: add AU CE integration tests
