+++ tests/ink-contracts/async-xcm-call-no-ce.wasm
