+++ tests/ink-contracts/call_xvm_payable.json
