+++ tests/ink-contracts/dummy_error.json
