+++ tests/ink-contracts/dummy_error.wasm
