+++ tests/ink-contracts/flipper.wasm
