+++ tests/ink-contracts/payable.wasm
