+++ tests/ink-contracts/simple_storage.wasm
