--- tests/xcm-simulator/fixtures/README.md
@@ -1,7 +1,0 @@
-## fixtures
-This directory contains contracts which are used in experimental scenarios.
-
-Files in this directory are used by experimental scenarios in `src/experimental.rs`. The json
-files are for informational purposes only and are not consumed by the tests.
-
-The source code for the contracts can be found at https://github.com/AstarNetwork/ink-test-contracts
